Resolve minor GCC warnings in various headers
In particular, these are the ones that show up when compiling
c_interface.cpp with g++ 4.5.2.

+++ src/cpp_interface/cpp_interface.h
@@ -63,8 +63,8 @@ namespace BEEV
 
     void checkInvariant()
     {
-      assert(bm.getAssertLevel() == cache.size());
-      assert(bm.getAssertLevel() == symbols.size());
+      assert(bm.getAssertLevel() == (int)cache.size());
+      assert(bm.getAssertLevel() == (int)symbols.size());
     }
 
     void init()
